lint

+++ contracts/interfaces/IComptroller.sol
@@ -2,8 +2,8 @@ pragma solidity 0.7.6;
 
 import "./IVBep20.sol";
 
-
 interface IComptroller {
     function getAllMarkets() external view returns (IVBep20[] memory);
-    function enterMarkets(address[] calldata vTokens) external returns (uint[] memory);
+
+    function enterMarkets(address[] calldata vTokens) external returns (uint256[] memory);
 }
